+++ openvino_xai/explainer/explanation.py
@@ -126,9 +126,10 @@ class Explanation:
         explain_target_indices = get_explain_target_indices(targets, label_names)
         if self.total_num_targets is not None:
             if not all(0 <= target_index <= (self.total_num_targets - 1) for target_index in explain_target_indices):
-                raise ValueError(
-                    f"All targets explanation indices have to be in range 0..{self.total_num_targets - 1}."
-                )
+                raise ValueError(f"All targets indices have to be in range 0..{self.total_num_targets - 1}.")
+        else:
+            if not all(target_index in self.saliency_map for target_index in explain_target_indices):
+                raise ValueError("Provided targer index {targer_index} is not available among saliency maps.")
         return explain_target_indices
 
     def save(self, dir_path: Path | str, name: str | None = None) -> None: