+++ CHANGELOG.md
@@ -4,11 +4,11 @@
 
 ### Summary
 
-*
+* Enable AISE: Adaptive Input Sampling for Explanation of Black-box Models.
 
 ### What's Changed
 
-*
+* Enable AISE: Adaptive Input Sampling for Explanation of Black-box Models by @negvet in https://github.com/intel-sandbox/openvino_xai/pull/49
 
 ### Known Issues